refactor: Split tests.py
Resolves: #7
